% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valohai-api.R
\name{generateUri}
\alias{generateUri}
\title{Generate uri for https requests}
\usage{
generateUri(uri = "", id)
}
\arguments{
\item{uri}{url for the valohai api}

\item{id}{optional id.}
}
\description{
This is a simple wrapper function which will take any valohai api, possibly with an id and generate the necessary uri for creating handles, posting etc.
This function is mainly intended for internal use, but can be used in case of multiple requests, to generate a handle that should be used for all requests which may lead to a slight performance boost.
}
\examples{
generateUri('executions')
generateUri('/executions////')
generateUri('executions', 3)

}
