% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valohai-api.R
\name{getAllUris}
\alias{getAllUris}
\title{Get a list of all uris available}
\usage{
getAllUris(token)
}
\arguments{
\item{token}{user token}
}
\description{
Simple function for importing a list of all available uris.
}
\examples{
\dontrun{
token <- 'YOUR-TOKEN-HERE'
getAllUris(token = token)
}

}
