% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valohai-api.R
\name{postValohai}
\alias{postValohai}
\title{POST requests for valohai}
\usage{
postValohai(token, uri, body, encoding = "json", id, ...)
}
\arguments{
\item{token}{user token}

\item{uri}{the uri parts}

\item{body}{data to be send with the body.}

\item{encoding}{the encoding format for the request. Defaults to json (application/json) but may also be multiform, form or raw. See help(httr::POST) for more information.}

\item{id}{execution, project or otherwise id}

\item{...}{additional arguments passed to httr::POST}
}
\value{
A list or tibble containing the named results.
}
\description{
This functions provides a basic interface to the valohai POST requests API.
The user it expected to provide a secret token, which should be able to read one or more urls. The available urls can be found using getAllUris() (see example).
It is important to note that if the 'uri' has multiple parts, it should be provided as a vector (eg 'executions/{id}/' should be given as c('executions', '{id}')), as the uri will be cleaned before being sent to the valohai API.
}
\examples{

\dontrun{
token <- 'YOUR-TOKEN-HERE'
postValohai(token,
 "executions",
 list(command = "command from valohai step"
     , commit = "commit id'"
     , environment = "environment id"
     , image = "rocker/rstudio:latest"
     , inherit_environment_variables = TRUE
     , inputs = list()
     , parameters = list("param name 1" = "hello world")
     , step = "download_yahoo"
     , tags = list("very useful tag for online gui and subsetting")
     , time_limit = 60 * 30 # Time limit.
     , project = "project id from getValohai(token, "environment")"
     , title = "R execution")
)
}

}
