% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valohai-api.R
\name{valohaiPrettify}
\alias{valohaiPrettify}
\title{Function for prettifying the result of any valohai call.}
\usage{
valohaiPrettify(x)
}
\arguments{
\item{x}{Any result from getValohai, postValohai, putValohai, etc.}
}
\value{
Possibly prettified result of x, given by the function in getOption('valohaiApiPrettifyFunc')
}
\description{
Function for prettifying the result of any valohai call.
}
